Add exemplar support
This simply works by using a fork of weaveworks/common, which
adds an exemplar to each instrumented HTTP request. In a real-life
application, only a fraction of requests should be used to add
exemplars. This can easily be made configurable, but is not yet in the
weaveworks/common fork. (See there for TODOs.)

For a demo, it's actually preferred that each requests sets an
exemplar.

Signed-off-by: beorn7 <beorn@grafana.com>