Use Grafana Agent for collecting metrics, logs, and traces
Use a DaemonSet of the Grafana Agent to collect metrics (using host filtering),
logs (mounting the node log directory), and traces (using a service for accepting
thrift, zipkin, and oltp traces).

This change configures the Prometheus instance to accept remote writes (from the
Grafana Agent) and not to scrape any pods (since this is handled by the agent).
It removes promtail (Loki client) and instead uses the promtail embedded in the
Grafana Agent.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>